requirements.txt
streamlit==1.3.0
transformers==4.12.2
Pillow==9.0.1
pandas==1.3.5
